[FEAT] Algebra plugins

+++ src/interfaces/IReflexRouter.sol
@@ -8,12 +8,9 @@ interface IReflexRouter {
     /// @param token0In Whether token0 is being swapped in.
     /// @param recipient The address to receive the profit.
     /// @return  profit The profit made from the backrun swap.
-    function triggerBackrun(
-        bytes32 triggerPoolId,
-        uint112 swapAmountIn,
-        bool token0In,
-        address recipient
-    ) external returns (uint profit, address profitToken);
+    function triggerBackrun(bytes32 triggerPoolId, uint112 swapAmountIn, bool token0In, address recipient)
+        external
+        returns (uint256 profit, address profitToken);
 
     /// @notice Returns the admin/owner address of the Reflex router.
     /// @return The address of the admin/owner.
